feat(archive): show banner for archived content
